- The withdrawal endpoint has been moved to RTK Query - Minor changes in translations - Significant code improvements: component decomposition, moving functions into hooks - Fixed a deposit bug - Increased icon sizes and raised the mobile navigation footer - Refactored userSlice.ts
