+++ public/tonconnect-manifest.json
@@ -1,8 +1,8 @@
 {
-  "url": "https://24tronex.ru",
-  "name": "TRONEX APP",
-  "iconUrl": "https://24tronex.ru/icon.png",
-  "termsOfUseUrl": "https://24tronex.ru/terms",
-  "privacyPolicyUrl": "https://24tronex.ru/privacy",
-  "redirectUrl": "https://24tronex.ru"
+  "url": "https://luckily-renewing-kookaburra.cloudpub.ru",
+  "name": "Your dApp name",
+  "iconUrl": "https://luckily-renewing-kookaburra.cloudpub.ru/icon.png",
+  "termsOfUseUrl": "https://luckily-renewing-kookaburra.cloudpub.ru/terms",
+  "privacyPolicyUrl": "https://luckily-renewing-kookaburra.cloudpub.ru/privacy",
+  "redirectUrl": "https://luckily-renewing-kookaburra.cloudpub.ru/"
 }
