+++ src/features/index.ts
@@ -4,6 +4,9 @@ export { UpgradeControl } from "./upgrade-level/ui/UpgradeControl.tsx";
 export { UpgradeTier } from "./upgrade/ui/UpgradeTier.tsx";
 export { BenefitCard } from "./bonus/ui/BenefitCard.tsx";
 export { WithdrawForm } from "./withdraw-form/ui/WithdrawForm.tsx";
+export { WithdrawAmountField } from "./withdraw-form/ui/WithdrawAmountField.tsx";
+export { WalletAddressField } from "./withdraw-form/ui/WalletAddressField.tsx";
+export { NetworkInfo } from "./withdraw-form/ui/NetworkInfo.tsx";
 export { DepositForm } from "./deposit-form/ui/DepositForm.tsx";
 export { LanguageSelector } from "./language-selector/ui/LanguageSelector.tsx";
