+++ src/pages/home/ui/HomePage.tsx
@@ -1,5 +1,5 @@
 import {
-  initDataRaw as _initDataRaw,
+  // initDataRaw as _initDataRaw,
   initDataState as _initDataState,
   useSignal,
 } from "@telegram-apps/sdk-react";
@@ -20,10 +20,11 @@ import { PATHS } from "@/shared/config/navigation.ts";
 import { isApiError } from "@/shared/lib";
 
 export const HomePage = () => {
-  const initDataRaw = useSignal(_initDataRaw);
+  // const initDataRaw = useSignal(_initDataRaw);
   const initDataState = useSignal(_initDataState);
   const navigate = useNavigate();
-
+  const initDataRaw =
+    "query_id=AAHTiAM1AAAAANOIAzXff-yb&user=%7B%22id%22%3A889424083%2C%22first_name%22%3A%22Pavel%22%2C%22last_name%22%3A%22%22%2C%22username%22%3A%22PaHuMbIu%22%2C%22language_code%22%3A%22ru%22%2C%22allows_write_to_pm%22%3Atrue%2C%22photo_url%22%3A%22https%3A%5C%2F%5C%2Ft.me%5C%2Fi%5C%2Fuserpic%5C%2F320%5C%2FLTql11elyHAZbfp37pqXmjxnJL_zMzjbqs1I4R0b2pA.svg%22%7D&auth_date=1757067236&signature=-XgWVCDQ5moypXW-TZdvlBPlwvN623sWKtjIv4yedFuAr8K5ym5ywJu6Q3cr1DxyEvV_nZuuLV0ITIRtFYqiBQ&hash=facbf33cf4925a55bd400b5a31a4bc7393a5a230bfeeb34e7d28fbf29db2d1a9";
   const [sendAuth] = useSendAuthMutation();
   const [getLevels] = useLazyGetLevelsQuery();
   const [getReferrals] = useLazyGetReferralsQuery();
