+++ src/shared/hooks/index.ts
@@ -2,3 +2,5 @@ export { useAppDispatch, useAppSelector } from "./useAppDispatch.ts";
 export { useIntersectionObserver } from "./useIntersectionObserver.ts";
 export { useProgressClick } from "./useProgressClick.ts";
 export { useRequiredContext } from "./useRequiredContext.ts";
+export { useDepositTransaction } from "./useDepositTransaction.ts";
+export { useWithdrawTransaction } from "./useWithdrawTransaction.ts";
