+++ src/shared/types/deposit.ts
@@ -1,0 +1,13 @@
+export interface DepositInfo {
+  id: number;
+  network: string;
+  amount: string;
+  status: string;
+  createdAt: string;
+}
+
+export interface DepositData {
+  page: number;
+  total: number;
+  deposits: DepositInfo[];
+}
